Updates Cases.py to use the same sort and display functions from Skins.py.
Updated Skins.py with find_case function that creates a list of cases an item can be inside.  Updated display function to display a list of cases for each matching item.  Updated get_item_color function to read the hex value from the JSON file and assign a color name for the display function to read